Fixing bug in evaluation of baseline
There were two issues previously:

1 - The evaluation simply checked if the template and tags were correct,
rather than comparing the merged SQL query.
2 - The template generation did not take into consideration cases where
a variable is not explicitly given, but is sentence specific (e.g.
afternoon).

Fixing these makes the task harder since there are now more templates,
but they also make it easier because the tags can have errors that do
not impact the query (e.g. a tag that does not occur in the template).